Move FIFO into storage.py so that proxies don't have to worry about it, and fix up HTTPS proxying

+++ storage.py
@@ -12,26 +12,47 @@ class StorageService:
 
 class AmazonSQS(StorageService):
     def __init__(self):
+        self.last_msg = -1
+        self.buffered = []
+        self.msg_num = 0
         self.access_key_id = ""
         self.secret_access_key = ""
         self.conn = SQSConnection(self.access_key_id, self.secret_access_key, True, None, None, None, None, None)
-    def put(self, q, data):
+    def put(self, q, data, use_msg_num = False):
         m = Message()
+        if (use_msg_num):
+            data = str(self.msg_num) + " " + data
+            self.msg_num += 1
         m.set_body(data)
-        print "Putting data: %s"%data
+#        print "Putting data: %s"%data
         status = q.write(m)
-        print "Status:"
-        print status
         if (status == False):
             print "Put failed"
             return False
         return True
-    def get(self, q):
+    def get(self, q, use_msg_num = False):
+        if (len(self.buffered) > 0 and self.buffered[0][0] == self.last_msg + 1):
+            data = self.buffered[0][1]
+            self.buffered = self.buffered[1:]
+            self.last_msg += 1
+            return data
         m = q.read()
         if (m == None):
             return False
         else:
             q.delete_message(m)
-            return m.get_body()
+            if (not use_msg_num):
+                return m.get_body()
+            else:
+                msg_num, sep, msg = m.get_body().partition(" ")
+                msg_num = int(msg_num)
+                if (msg_num == self.last_msg + 1):
+#                    print "Using msg %d"%msg_num
+                    self.last_msg = msg_num
+                    return msg
+                else:
+#                    print "Buffering msg %d, last_msg: %d"%(msg_num, self.last_msg)
+                    self.buffered.append((msg_num, msg))
+                    self.buffered.sort()
     def create_queue(self, name):
         return self.conn.create_queue(name)
